\name{compute_sufficient_statistics}
\alias{compute_sufficient_statistics}
\title{
Compute sufficient statistics
}
\description{
Compute sufficient statistics given \eqn{y} and \eqn{z}.
}
\usage{
compute_sufficient_statistics(y, z, K, x_data)
}
\arguments{
  \item{y}{
	Matrix of factors
}
  \item{z}{
	Allocation vector
}
  \item{K}{
	Number of components
}
  \item{x_data}{
	Data
}
}
\value{
	A list with six entries of sufficient statistics.
}
\author{
Panagiotis Papastamoulis
}

