\name{compute_sufficient_statistics_q0}
\alias{compute_sufficient_statistics_q0}
\title{
Compute sufficient statistics for \eqn{q = 0}
}
\description{
Compute sufficient statistics given \eqn{z}.
}
\usage{
compute_sufficient_statistics_q0(z, K, x_data)
}
\arguments{
  \item{z}{
	Allocation vector
}
  \item{K}{
	Number of components
}
  \item{x_data}{
	Data
}
}
\value{
	A list with six entries of sufficient statistics.
}
\author{
Panagiotis Papastamoulis
}

