\name{fabMix-package}
\alias{fabMix-package}
\alias{fabMix}
\docType{package}
\title{
\packageTitle{fabMix}
}
\description{
\packageDescription{fabMix}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{fabMix}
\packageIndices{fabMix}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{fabMix}

Maintainer: \packageMaintainer{fabMix}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
