\name{log_dirichlet_pdf}
\alias{log_dirichlet_pdf}
\title{
Log-density function of the Dirichlet distribution
}
\description{
Log-density function of the Dirichlet distribution
}
\usage{
log_dirichlet_pdf(alpha, weights)
}
\arguments{
  \item{alpha}{
	Parameter vector
}
  \item{weights}{
	Vector of weights
}
}
\value{
	Log-density of the \eqn{D(alpha_1,\ldots,\alpha_k)} evaluated at \eqn{w_1,\ldots,w_k}.
}
\author{
Panagiotis Papastamoulis
}


