\name{myDirichlet}
\alias{myDirichlet}
\title{
Simulate from the Dirichlet distribution
}
\description{
Generate a random draw from the Dirichlet distribution \eqn{D(\alpha_1,\ldots,\alpha_k)}.
}
\usage{
myDirichlet(alpha)
}
\arguments{
  \item{alpha}{
	Parameter vector
}
}
\value{
	Simulated vector
}
\author{
Panagiotis Papastamoulis
}

