\name{readLambdaValues}
\alias{readLambdaValues}
\title{
Read Lambda values.
}
\description{
Function to read Lambda values from file. 
}
\usage{
readLambdaValues(myFile,K,p,q)
}
\arguments{
  \item{myFile}{
	File containing Lambda values
}
  \item{K}{
	Number of components
}
  \item{p}{
	Number of variables
}
  \item{q}{
	Number of factors
}
}
\value{
	\eqn{K\times p \times q} array of factor loadings. 
}
\author{
Panagiotis Papastamoulis
}

