\name{update_OmegaINV_Cxx}
\alias{update_OmegaINV_Cxx}
\title{
Gibbs sampling for \eqn{\Omega^{-1}} for Cxx model
}
\description{
Gibbs sampling for \eqn{\Omega^{-1}} for Cxx model
}
\usage{
update_OmegaINV_Cxx(Lambda, K, g, h)
}
\arguments{
  \item{Lambda}{
	Factor loadings
}
  \item{K}{
	Number of components
}
  \item{g}{
	Prior parameter
}
  \item{h}{
	Prior parameter
}
}
\value{
	\eqn{\Omega^{-1}}
}
\author{
Panagiotis Papastamoulis
}

