    \name{waveDataset1500}
     \docType{data}
     \alias{waveDataset1500}
     \title{Wave dataset}
     \description{A subset of 1500 randomly sampled observations from the wave dataset (version 1), available from the UCI machine learning repository. It contains 3 classes of waves (variable \code{class} with values ``1'', ``2'' and ``3'') and 21 attributes. Each class is generated from a combination of 2 of 3 base waves with noise.}
     \usage{data(waveDataset1500)}
     \format{A data frame with 1500 rows and 22 columns. The first column denotes the class of each observation.}
     \keyword{datasets}
     \source{\url{https://archive.ics.uci.edu/ml/datasets/Waveform+Database+Generator+(Version+1)}}
     \references{
Lichman, M. (2013). UCI Machine Learning Repository \url{http://archive.ics.uci.edu/ml}. Irvine, CA: University of California, School of Information and Computer Science.

Breiman,L., Friedman,J.H., Olshen,R.A. and Stone,C.J. (1984). Classification and Regression Trees. Wadsworth International Group: Belmont, California.
}

