

\name{print.fabMix.object}
\alias{print.fabMix.object}
\title{
Print function
}
\description{
This function prints a summary of objects returned by the \code{fabMix} function.
}
\usage{
\method{print}{fabMix.object}(x, ...)
}
\arguments{
  \item{x}{
An object of class \code{fabMix.object}, which is returned by the \code{fabMix} function.
}
  \item{...}{
        ignored.
}
}
\details{
The function prints some basic information for a \code{fabMix.object}.
}
\author{
Panagiotis Papastamoulis
}

